+++ src/main/java/roomescape/reservation/presentation/AdminReservationController.java
@@ -3,9 +3,13 @@ package roomescape.reservation.presentation;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.member.application.MemberService;
 import roomescape.member.domain.Member;
@@ -17,6 +21,9 @@ import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.dto.request.AdminReservationSaveRequest;
 import roomescape.reservation.dto.response.ReservationResponse;
+
+import java.time.LocalDate;
+import java.util.List;
 
 @RestController
 @RequestMapping("/admin/reservations")
@@ -43,5 +50,30 @@ public class AdminReservationController {
         Reservation createdReservation = reservationService.create(newReservation);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ReservationResponse.from(createdReservation));
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ReservationResponse>> findReservations() {
+        List<Reservation> reservations = reservationService.findAll();
+        return ResponseEntity.ok(reservations.stream()
+                .map(ReservationResponse::from)
+                .toList());
+    }
+
+    @GetMapping("/searching")
+    public ResponseEntity<List<ReservationResponse>> findReservationsByMemberIdAndThemeIdAndDateBetween(
+            @RequestParam Long memberId, @RequestParam Long themeId,
+            @RequestParam LocalDate fromDate, @RequestParam LocalDate toDate) {
+        List<Reservation> reservations = reservationService.findReservationsByMemberIdAndThemeIdAndDateBetween(
+                memberId, themeId, fromDate, toDate);
+        return ResponseEntity.ok(reservations.stream()
+                .map(ReservationResponse::from)
+                .toList());
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        reservationService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
