+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.member.domain.Member;
 import roomescape.reservation.application.ReservationService;
@@ -22,7 +21,6 @@ import roomescape.reservation.dto.request.ReservationSaveRequest;
 import roomescape.reservation.dto.response.MyReservationResponse;
 import roomescape.reservation.dto.response.ReservationResponse;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -50,31 +48,6 @@ public class ReservationController {
         Reservation createReservation = reservationService.create(newReservation);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(ReservationResponse.from(createReservation));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationResponse>> findReservations() {
-        List<Reservation> reservations = reservationService.findAll();
-        return ResponseEntity.ok(reservations.stream()
-                .map(ReservationResponse::from)
-                .toList());
-    }
-
-    @GetMapping("/searching")
-    public ResponseEntity<List<ReservationResponse>> findReservationsByMemberIdAndThemeIdAndDateBetween(
-            @RequestParam Long memberId, @RequestParam Long themeId,
-            @RequestParam LocalDate fromDate, @RequestParam LocalDate toDate) {
-        List<Reservation> reservations = reservationService.findReservationsByMemberIdAndThemeIdAndDateBetween(
-                memberId, themeId, fromDate, toDate);
-        return ResponseEntity.ok(reservations.stream()
-                .map(ReservationResponse::from)
-                .toList());
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.delete(id);
-        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/mine")
